[OPS-1365] Add advanced benchwrapper CPU management
Problem: Currently, 'benchwrapper' unconditionally reserves last 'n'
CPUs for the given task. As a result, when two 'benchwrapper's are
running, their tasks share CPUs which is a problem when these task are
some benchmarks that are supposed to be run in an isolated environment.

Solution: Add advanced CPU management. 'benchwrapper' now gets the set
of available CPUs and reserves required amount of CPUs from it
returning them back to the system pool once the given task is done.