[OPS-1468] Add function to parse ghc versions from cabal file
Problem: We have a CI template for libraries which suggests specifying
it in ghc-versions. So if we use this template and specify
tested-with, we duplicate this information. Specifying it in
tested-with seems better, because this is a more standardized way and
is more accessible by users of the package.

Solution: Add function to parse ghc versions from cabal file.