acme-sh: install cert to a dedicated directory
Problem: acme.sh changed the default directory to which certiticate is
initially issued. As a result, vault doesn't use newly-fetched
certificate. Additionally, acme.sh README says that one must copy
certificate to a desired location instead of relying on the structure of
acme.sh state directory.

Solution: '--install-cert' to a dedicated directory and point 'keyPath'
and 'certPath' to it.