[OPS-1458] Add haskell.nix wrapper for CI
Problem: Our CI template for libraries allows one to specify
ghc-versions so that we can build a library in different
configurations. Currently the only configurable thing is the GHC
version. We also want to be able to build libraries with different
stack.yaml files and stack resolvers.

Solution: Add a haskell.nix wrapper to the library CI capable of
building a project with different GHC versions, stack.yaml files, and
stack resolvers.