Add module with oneshot service for postgresql migration
Problem: From time to time we need to update the version of postgresql
running on our infra. When a major version of postgres is changed,
usually, a dedicated migration for data directory is needed.

Solution: To remove the necessity to do the migration manually,
introduce a module that will add a oneshot service which will perform a
migration prior to starting postresql service.